+++ include/logdest.h
@@ -20,8 +20,8 @@ struct _logdest {
 typedef struct _logdest logdest_t;
 struct _logres;
 
-extern int logdest_get_arg(const uint8_t *buf, uint32_t buflen, uint8_t type, uint8_t argn, 
-                const void** ptr, uint32_t *len, uint8_t *type_out, uint8_t *arg_out);
+extern int logdest_get_arg(const uint8_t *buf, uint32_t buflen, uint8_t type, uint32_t argn, 
+                const void** ptr, uint32_t *len, uint8_t *type_out, uint32_t *arg_out);
 extern void logdest_format_message_stream(void (*putstr)(void *out_buf, const char *str), void *out_buf, struct _logres *res,
                 const char *fmt, const uint8_t *buf, uint32_t len);
 extern void logdest_format_message(struct _logres *res, const char *fmt, const uint8_t *buf, uint32_t len, char **out);
